cambiar el nombre de algunos archivos y eliminar componentes no utilizables
